Check if the device exists before adding a value

+++ device_key_value.py
@@ -7,6 +7,9 @@ from google.appengine.ext import ndb
 
 import webapp2
 
+class Device(ndb.Model):
+    device_id = ndb.StringProperty()
+
 class Value(ndb.Model):
     device_id = ndb.StringProperty()
     device_key = ndb.StringProperty()
@@ -28,17 +31,23 @@ class SetPage(webapp2.RequestHandler):
                 self.response.write('ERROR')
 
         else:
-            q = Value.query(Value.device_id == device_id, Value.device_key == device_key)
+            q = Device.query(Device.device_id == device_id)
 
-            for p in q:
-                p.key.delete()
+            if q.count() != 0:
 
-            value = Value()
-            value.device_id = self.request.get('device')
-            value.device_key = self.request.get('key')
-            value.device_value = self.request.get('value')
-            value.put()
-            self.response.write('OK')
+                q = Value.query(Value.device_id == device_id, Value.device_key == device_key)
+
+                for p in q:
+                    p.key.delete()
+
+                value = Value()
+                value.device_id = self.request.get('device')
+                value.device_key = self.request.get('key')
+                value.device_value = self.request.get('value')
+                value.put()
+                self.response.write('OK')
+            else:
+                self.response.write('ERROR')
 
 app = webapp2.WSGIApplication([
     ('/', SetPage),
